1) In the material_nums table, I set all the prices to 0. 2) I set some of the material_nums prices based on pre-existing commodities. 3) I modified the mob loading code to use the prices set above when loading commods, rather than referring to the commodity object. 4) I added a TObj::suggestedPrice function but left it disabled. 5) I added calls to TObj::suggestedPrice to all of the derived suggestedPrice functions. 6) I modified the shop code for commodities to use the material_num prices rather than the wackass way it was doing it before.
http://forums.sneezymud.com/mudadmin/smf/index.php/topic,4975.0.html

+++ code/create_objs.cc
@@ -3229,7 +3229,7 @@ void TTrap::changeTrapValue3(TBeing *ch, const char *arg, editorEnterTypeT type)
   ch->sendTo("Select the number to set to, <ENTER> to return to main menu.\n\r--> ");
 }
 
-void TObj::describeTreasure(const char *arg,int num, int price)
+void TObj::describeTreasure(const char *arg,int num, float price)
 {
   char type[80];
   char buf[160];
@@ -3277,7 +3277,7 @@ void TObj::describeTreasure(const char *arg,int num, int price)
   setDescr(mud_str_dup(buf));
   
   //  set value and rent 
-  obj_flags.cost = num * price;
+  obj_flags.cost = (int)(num * price);
 
   setWeight((float) (num + 0.5)/10.0);
 }
