+++ code/spec_mobs.cc
@@ -6049,7 +6049,7 @@ int commodMaker(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *o
     commod=read_object(material_nums[v->getMaterial()].availability, VIRTUAL);
 
     commod->describeTreasure(material_nums[v->getMaterial()].mat_name,
-		      v->obj_flags.cost/material_nums[v->getMaterial()].price,
+		      (int)(v->obj_flags.cost/material_nums[v->getMaterial()].price),
 			     material_nums[v->getMaterial()].price);
 
     me->doSay("Alright, here you go!");
